Updated Barchart to Shruti world map
Added a GDP Barchart to Shruti's world Map. World Map interacts with Barchart. This Barchart shows top 10 countries depending on the continent selected in the dropdown. Bar chart interacts with world map when you hover on a particular country bar. Please feel free to suggest if you have any ideas on colors and cosmetics for Barchart or button.

Also added a button to reset the world map to its original look after the interaction is complete.